requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.35
pymysql==1.1.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
crewai==0.51.1
langchain-openai==0.2.0
arxiv==2.1.3
requests==2.32.3